requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SocketIO
SQLAlchemy
python-dotenv
Jinja2
itsdangerous
eventlet
gunicorn
email-validator
Werkzeug
click
bleach
slugify
libsass
Flask-Assets
WTForms
Flask-WTF
